feat: remove specific farem applications
