--- apps/farem-go/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "farem-go"
-version = "0.1.0"
-edition = "2021"
-description = "Responsible for the Go farem"
-
-[dependencies]
-duct = "0.13.5"
-rocket = { version = "0.5.0-rc.2", features = ["json"] }
-serde = { version = "1.0.144", features = ["derive"] }
-utilities = { path = "../../libs/utilities" }
