--- apps/farem-go/project.json
@@ -1,38 +1,0 @@
-{
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "implicitDependencies": ["examples"],
-  "projectType": "application",
-  "sourceRoot": "apps/farem-go/src",
-  "targets": {
-    "build": {
-      "executor": "@ignisda/nx-rust:build",
-      "outputs": [
-        "target/debug/farem-go*",
-        "target/debug/*/farem-go*",
-        "target/release/farem-go*",
-        "target/release/*/farem-go*"
-      ],
-      "options": {
-        "release": false
-      },
-      "configurations": {
-        "production": {
-          "release": true
-        }
-      }
-    },
-    "test": {
-      "executor": "@ignisda/nx-rust:nextest",
-      "options": {}
-    },
-    "lint": {
-      "executor": "@ignisda/nx-rust:clippy",
-      "options": {
-        "fix": false,
-        "failOnWarnings": true,
-        "noDeps": true
-      }
-    }
-  },
-  "tags": []
-}
