+++ workspace.json
@@ -7,9 +7,6 @@
     "config": "libs/config",
     "domains": "libs/domains",
     "examples": "libs/examples",
-    "farem-cpp": "apps/farem-cpp",
-    "farem-go": "apps/farem-go",
-    "farem-rs": "apps/farem-rs",
     "farem-run": "apps/farem-run",
     "generated": "libs/generated",
     "graphql": "libs/graphql",
